requirements.txt
altair==4.2.0
pandas==1.5.2
PyMySQL==1.0.2
PyYAML==6.0
streamlit==1.15.2
cryptography
